+++ src/DraxSubprovider.tsx
@@ -2,10 +2,10 @@ import React, { FunctionComponent } from 'react';
 
 import { DraxContext } from './DraxContext';
 import { DraxSubproviderProps } from './types';
-import { useDrax } from './hooks';
+import { useDraxContext } from './hooks';
 
 export const DraxSubprovider: FunctionComponent<DraxSubproviderProps> = ({ parent, children }) => {
-	const contextValue = useDrax();
+	const contextValue = useDraxContext();
 	const subContextValue = {
 		...contextValue,
 		parent,