+++ src/hooks/index.ts
@@ -1,4 +1,4 @@
-export { useDrax } from './useDrax';
+export { useDraxContext } from './useDraxContext';
 export { useDraxId } from './useDraxId';
 export { useDraxRegistry } from './useDraxRegistry';
 export { useDraxState } from './useDraxState';