+++ src/hooks/useDraxContext.ts
@@ -2,7 +2,7 @@ import { useContext } from 'react';
 
 import { DraxContext } from '../DraxContext';
 
-export const useDrax = () => {
+export const useDraxContext = () => {
 	const drax = useContext(DraxContext);
 	if (!drax) {
 		throw Error('No DraxProvider found');